+++ geotrek/zoning/models.py
@@ -11,7 +11,7 @@ from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
 from geotrek.common.utils import uniquify
-from geotrek.core.models import Topology, Path, Trail
+from geotrek.core.models import Topology, Path
 from geotrek.maintenance.models import Intervention, Project
 
 
@@ -74,8 +74,12 @@ class RestrictedAreaEdge(Topology):
 
     @classmethod
     def topology_area_edges(cls, topology):
-        return cls.overlapping(topology).select_related('restricted_area')\
-                                        .select_related('restricted_area__area_type')
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            qs = cls.overlapping(topology)
+        else:
+            qs = cls.objects.filter(geom__intersects=topology.geom)
+        return qs.select_related('restricted_area')\
+                 .select_related('restricted_area__area_type')
 
 Path.add_property('area_edges', RestrictedAreaEdge.path_area_edges)
 Path.add_property('areas', lambda self: uniquify(map(attrgetter('restricted_area'), self.area_edges)))
@@ -128,7 +132,11 @@ class CityEdge(Topology):
 
     @classmethod
     def topology_city_edges(cls, topology):
-        return cls.overlapping(topology).select_related('city')
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            qs = cls.overlapping(topology)
+        else:
+            qs = cls.objects.filter(geom__intersects=topology.geom)
+        return qs.select_related('city')
 
 Path.add_property('city_edges', CityEdge.path_city_edges)
 Path.add_property('cities', lambda self: uniquify(map(attrgetter('city'), self.city_edges)))
@@ -179,7 +187,11 @@ class DistrictEdge(Topology):
 
     @classmethod
     def topology_district_edges(cls, topology):
-        return cls.overlapping(topology).select_related('district')
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            qs = cls.overlapping(topology)
+        else:
+            qs = cls.objects.filter(geom__intersects=topology.geom)
+        return qs.select_related('district')
 
 Path.add_property('district_edges', DistrictEdge.path_district_edges)
 Path.add_property('districts', lambda self: uniquify(map(attrgetter('district'), self.district_edges)))